add markdown stories for each package
