+++ docs/getting-started.stories.mdx
@@ -6,8 +6,6 @@ import { Meta } from "@storybook/addon-docs"
 
 The avail design system is made up of small UI packages that can be mostly imported into projects as needed. This allows updating a package to happen without worrying about breaking changes that might be in development on another components.
 
----
-
 | Package                  | Exports                                             |
 | ------------------------ | --------------------------------------------------- |
 | @rent_avail/autocomplete | Address Autocomplete                                |
