+++ packages/autocomplete/stories/_autocomplete.stories.mdx
@@ -1,11 +1,8 @@
 import { Meta, ArgsTable, Props, Story } from "@storybook/addon-docs"
 import { ManualOption, DefaultValue } from "./autocomplete.stories"
-import { Autocomplete, useAutocomplete, AutocompleteProvider } from "./src"
 
 <Meta title="Packages / Autocomplete" />
 
 <Story story={DefaultValue} />
 
 <Story story={ManualOption} />
-
-<Props of={Autocomplete} />
