+++ packages/avatar/stories/_avatar.stories.mdx
