+++ packages/core/stories/_core.stories.mdx
@@ -1,10 +1,8 @@
----
-title: Layout
-layout: elements
-package_name: layout
-install: "yarn add @rent_avail/core"
-source: "https://github.com/rentalutions/elements/tree/release/packages/layout"
----
+import { Meta, Story } from "@storybook/addon-docs"
+
+<Meta title="Packages / Core" />
+
+# Core
 
 ```js
 import { Box, Card, Container, Flex, Grid, Col, Stack } from "@rent_avail/core"
