+++ packages/dialog/stories/_dialog.stories.mdx
@@ -1,10 +1,8 @@
----
-title: Dialogs
-layout: elements
-package_name: dialog
-install: "yarn add @rent_avail/dialog"
-source: "https://github.com/rentalutions/elements/tree/release/packages/dialog"
----
+import { Meta, Story } from "@storybook/addon-docs"
+
+<Meta title="Packages / Dialog" />
+
+# Dialog
 
 ```js
 import {
