+++ packages/dialog/stories/dialog.stories.js
@@ -9,7 +9,7 @@ import {
   FullscreenDialog,
 } from "../src"
 
-export default { title: "Packages/Dialog" }
+// export default { title: "Packages/Dialog" }
 
 export function ConfirmationUsage() {
   const [open, set] = useState(false)
