+++ packages/feedback/stories/fullpage.story.tsx
@@ -1,13 +1,11 @@
-import React, { useState, Fragment } from "react"
-import { FullscreenFeedback, InlineFeedback } from "../src"
+import { useState } from "react"
 import { Avatar } from "@rent_avail/avatar"
 import { Box, Container, Card, Stack, Grid, Col, Flex } from "@rent_avail/core"
 import { Button, Input } from "@rent_avail/controls"
-import { Dialog, DialogTarget, ConfirmationDialog } from "@rent_avail/dialog"
-import { motion } from "framer-motion"
-import { Plus } from "react-feather"
+import { Plus } from "@rent_avail/icons"
+import { FullscreenFeedback } from "../src"
 
-export default { title: "Packages/Feedback" }
+// export default { title: "Packages/Feedback" }
 
 export function FullpagePayments() {
   const [open, setOpen] = useState(false)
@@ -225,180 +223,5 @@ export function FullpageSendLease() {
         }
       />
     </Container>
-  )
-}
-
-export function InlineVerification() {
-  const [loaded, setLoaded] = useState(false)
-  return (
-    <Container
-      as={Grid}
-      minHeight="calc(100vh - 8rem)"
-      mt="4rem"
-      display="grid"
-      alignItems="center"
-    >
-      <Col minHeight="32rem" span={[12, 10]} offset={["auto", 2]}>
-        <InlineFeedback
-          steps={[
-            "Verifying Phone",
-            "Generating Device Data",
-            "Securing Account",
-          ]}
-          onAnimationEnd={() => setLoaded(true)}
-        />
-        {loaded && (
-          <Box
-            as={motion.section}
-            mt="2rem"
-            initial={{ opacity: 0, scale: 0.9 }}
-            animate={{ opacity: 1, scale: 1 }}
-          >
-            <Box as="h3" mb="2rem">
-              Great, your account is secured.
-            </Box>
-            <Box>You're all set to start using your Avail account.</Box>
-            <Box my="2rem" p="2rem" bg="blue_100" borderRadius="0.25rem">
-              <Box fontWeight="bold">Account Security</Box>
-              <Box>
-                The next time you log in from a new device, you'll recive an
-                authentication code for this number to keep your account secure.
-              </Box>
-            </Box>
-            <Box display="flex" justifyContent="flex-end" mt="2rem">
-              <Button variant="primary">Continue</Button>
-            </Box>
-          </Box>
-        )}
-      </Col>
-    </Container>
-  )
-}
-
-export function InlineTransunion() {
-  const [loaded, setLoaded] = useState(false)
-  const reportProps = {
-    bg: "blue_100",
-    borderRadius: "0.25rem",
-    display: "flex",
-    alignItems: "center",
-    p: "0.25rem",
-  }
-  const reportBoxProps = {
-    as: "h3",
-    flexBasis: "50%",
-    bg: "ui_100",
-    p: "2rem",
-  }
-  const reportLabelProps = {
-    as: "h5",
-    p: "2rem",
-  }
-  return (
-    <Container mt="4rem">
-      <Card>
-        <Box
-          as="header"
-          borderBottom="1px solid"
-          borderBottomColor="ui_500"
-          mb="2rem"
-          pb="2rem"
-        >
-          <Box>Application Details</Box>
-        </Box>
-        <InlineFeedback
-          steps={[
-            "Checking Data",
-            "Compiling Reports",
-            "Verifying Identity",
-            "Fetching Score",
-          ]}
-          onAnimationEnd={() => setLoaded(true)}
-        />
-        <Box mt="2rem">Screening Reports</Box>
-        {loaded && (
-          <Box
-            as={motion.section}
-            initial={{ opacity: 0, y: "1rem" }}
-            animate={{ opacity: 1, y: 0 }}
-            display="grid"
-            gridTemplateColumns={["1fr", "1fr", "repeat(3, 1fr)"]}
-            mt="2rem"
-            gridGap="2rem"
-          >
-            <Box {...reportProps}>
-              <Box {...reportBoxProps}>720</Box>
-              <Box {...reportLabelProps}>Credit Score</Box>
-            </Box>
-            <Box {...reportProps}>
-              <Box {...reportBoxProps}>0</Box>
-              <Box {...reportLabelProps}>Criminal Records</Box>
-            </Box>
-            <Box {...reportProps}>
-              <Box {...reportBoxProps}>0</Box>
-              <Box {...reportLabelProps}>Eviction Records</Box>
-            </Box>
-            <Box gridColumn="span 3" justifySelf="end">
-              <Button variant="primary">submit reports</Button>
-            </Box>
-          </Box>
-        )}
-      </Card>
-    </Container>
-  )
-}
-
-export function InlinePublishListing() {
-  const [open, setOpen] = useState(false)
-  const [loaded, setLoaded] = useState(false)
-  function handleToggle() {
-    setOpen(false)
-    setLoaded(false)
-  }
-  return (
-    <Box
-      display="grid"
-      gridTemplateColumns="25rem auto"
-      gridTemplateRows="15rem auto"
-      minHeight="calc(100vh - 4rem)"
-    >
-      <Box gridRow="span 2" bg="blue_100" />
-      <Box bg="blue_500" />
-      <Box bg="ui_300">
-        <Container mt="4rem">
-          <Flex justifyContent="flex-end">
-            <Dialog open={open} toggle={handleToggle} id="success-listing">
-              <DialogTarget>
-                <Button onClick={(e) => setOpen(true)}>Publish Listing</Button>
-              </DialogTarget>
-              <ConfirmationDialog>
-                <InlineFeedback
-                  steps={[
-                    "Optimizing Photos",
-                    "Connecting to Syndicates",
-                    "Updating Feed",
-                  ]}
-                  onAnimationEnd={() => setLoaded(true)}
-                />
-                {loaded && (
-                  <Box
-                    as={motion.div}
-                    initial={{ opacity: 0, y: "1rem" }}
-                    animate={{ opacity: 1, y: 0 }}
-                    mt="2rem"
-                  >
-                    <Box as="h3">Listing Published</Box>
-                    <Box mt="2rem">
-                      Successfully published to 12 sites, it can take up to 4
-                      hours for your listing to appear on our partner sites.
-                    </Box>
-                  </Box>
-                )}
-              </ConfirmationDialog>
-            </Dialog>
-          </Flex>
-        </Container>
-      </Box>
-    </Box>
   )
 }
