+++ packages/icons/stories/_icons.stories.mdx
@@ -1,0 +1,8 @@
+import { Meta, Story } from "@storybook/addon-docs"
+import { Icons } from "./default.story"
+
+<Meta title="Packages / Icons" />
+
+# Icons
+
+<Story story={Icons} />
