+++ packages/menu/stories/_menu.stories.mdx
@@ -1,0 +1,10 @@
+import { Meta, Story } from "@storybook/addon-docs"
+import BasicUsage from "./basic.story"
+import PortalUsage from "./portal.story"
+import ParentRefUsage from "./parent.story"
+
+<Meta title="Packages / Menu" />
+
+# Menu
+
+<Story story={BasicUsage} />
