--- packages/menu/stories/menu.stories.js
@@ -1,6 +1,0 @@
-import BasicUsage from "./basic"
-import PortalUsage from "./portal"
-import ParentRefUsage from "./parent-ref"
-
-export default { title: "Packages/Menu" }
-export { BasicUsage, PortalUsage, ParentRefUsage }
