+++ packages/popover/stories/_popover.stories.mdx
@@ -1,0 +1,5 @@
+import { Meta, Story } from "@storybook/addon-docs"
+
+<Meta title="Packages / Popover" />
+
+# Popover
