+++ packages/progress/stories/_progress.stories.mdx
@@ -1,0 +1,64 @@
+import { Meta, Story } from "@storybook/addon-docs"
+import { Default } from "./default.story"
+
+<Meta title="Packages / Progress" />
+
+# Progress
+
+<Story story={Default} />
+
+```js
+import { FixedProgress, FluidProgress } from "@rent_avail/progress"
+```
+
+A progress bar indicates the status of a user inside a process. It help a user understand what they've accomplished and what is yet to come. This gives them a realistic idea of time to completion and lowers uncertainty.
+
+## FixedProgress
+
+```jsx live
+<FixedProgress
+  steps={{
+    "Report Submitted": true,
+    "Rent Verified": false,
+    "Tradeline Created": false,
+  }}
+/>
+```
+
+### Description
+
+Fixed progress bars should be used to indicate a short workflow, less than 4 steps, or to denote the status of a workflow that will take an extended amount of time.
+
+#### Style Props
+
+`space`
+
+#### Props
+
+| Name  | Type     | Description                                                                                   |
+| ----- | -------- | --------------------------------------------------------------------------------------------- |
+| steps | `object` | An object with where the key is the step label and the value is whether the step is complete. |
+
+---
+
+## FluidProgress
+
+```jsx live
+<FluidProgress label="Contact info" current={2} total={6} />
+```
+
+### Description
+
+Fluid progress bar should be used in a user journey that has more than 4 steps. Each step in the workflow should be on it's own screen, functioning as a wizard.
+
+#### Style Props
+
+`space`
+
+#### Props
+
+| Name    | Type     | Description                                                                         |
+| ------- | -------- | ----------------------------------------------------------------------------------- |
+| label   | `object` | The label currently displayed on the progressBar, should be updated on each render. |
+| current | `number` | The index of the current step.                                                      |
+| total   | `number` | The length of the total steps.                                                      |
