+++ packages/progress/stories/default.story.tsx
@@ -1,0 +1,5 @@
+import { Box } from "@rent_avail/core"
+
+export function Default() {
+  return <Box>Progress Bar</Box>
+}
