--- packages/progress/stories/progess.stories.js
@@ -1,5 +1,0 @@
-export default { title: "Packages/Progress Bar" }
-
-export function Default() {
-  return "Hello progress bar."
-}
