+++ packages/skeleton/stories/_skeleton.stories.mdx
@@ -1,0 +1,5 @@
+import { Meta, Story } from "@storybook/addon-docs"
+
+<Meta title="Packages / Skeleton" />
+
+# Skeleton Loader
