+++ packages/toast/stories/_toast.stories.mdx
@@ -1,0 +1,18 @@
+import { Meta, Story } from "@storybook/addon-docs"
+import { ToastProvider } from "../src"
+import { BasicToast } from "./basic.story"
+
+<Meta
+  title="Packages / Toast"
+  decorators={[
+    (Story) => (
+      <ToastProvider>
+        <Story />
+      </ToastProvider>
+    ),
+  ]}
+/>
+
+# Toast
+
+<Story story={BasicToast} />
