+++ packages/tooltip/stories/tooltip.stories.js
@@ -9,7 +9,7 @@ import {
 import { Button } from "@rent_avail/controls"
 import { Tooltip, TooltipContent, TooltipTarget } from "../src"
 
-export default { title: "Packages/Tooltip" }
+// export default { title: "Packages/Tooltip" }
 
 export function BasicUsage() {
   return (
