+++ scripts/create-package.js
@@ -54,8 +54,8 @@ if (!packages.includes(name)) {
   cp("-r", "template", `packages/${name}`)
   writeFileSync(`packages/${name}/package.json`, packageJson)
   renameSync(
-    `packages/${name}/stories/package.stories.tsx`,
-    `packages/${name}/stories/${name}.stories.tsx`
+    `packages/${name}/stories/_package.stories.mdx`,
+    `packages/${name}/stories/_${name}.stories.mdx`
   )
   renameSync(
     `packages/${name}/__tests__/package.tsx`,
