+++ template/stories/_package.stories.mdx
@@ -1,0 +1,1 @@
+import { Meta, Story } from "@storybook/addon-docs"
